Refactor read loops (#39)
Make them read multiple chunks if possible. Also
fix a possible infinite loop if the input buffer is zero-length.

Unify all the check*State functions into one for encoders,
and another for decoders.

Error if an empty block is encountered in V2, except if the
entire message is empty.

Also add comment for assertEndOfStream().